requirements.txt
glob2==0.7
pybase64==1.2.1
pycryptodome==3.14.1
